+++ src/gui/guifx/TelaCadastroMedi.fxml
@@ -1,51 +1,21 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
 <?import javafx.scene.control.PasswordField?>
 <?import javafx.scene.control.Separator?>
 <?import javafx.scene.control.TextField?>
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.text.Font?>
-<?import javafx.scene.text.Text?>
 
-<AnchorPane id="AnchorPane" prefHeight="500.0" prefWidth="440.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="gui.guifx.TelaCadastroMedController">
+<AnchorPane id="AnchorPane" prefHeight="500.0" prefWidth="440.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="gui.guifx.TelaCadastroMediController">
    <children>
       <Separator layoutX="57.0" layoutY="246.0" prefHeight="6.0" prefWidth="327.0" />
-      <Text fx:id="titulo" layoutX="117.0" layoutY="65.0" strokeWidth="0.0" text="Cadastrar Usuários">
-         <font>
-            <Font size="24.0" />
-         </font>
-      </Text>
       <Separator layoutX="57.0" layoutY="299.0" prefHeight="6.0" prefWidth="327.0" />
-      <Text fx:id="textRecp" layoutX="62.0" layoutY="118.0" strokeWidth="0.0" text="Médico">
-         <font>
-            <Font size="16.0" />
-         </font>
-      </Text>
-      <Text fx:id="textCpf" layoutX="57.0" layoutY="226.0" strokeWidth="0.0" text="CPF: ">
-         <font>
-            <Font size="16.0" />
-         </font>
-      </Text>
       <TextField fx:id="filedCpf" layoutX="125.0" layoutY="207.0" prefHeight="25.0" prefWidth="254.0" />
-      <Text fx:id="textIdade" layoutX="59.0" layoutY="284.0" strokeWidth="0.0" text="Idade: ">
-         <font>
-            <Font size="16.0" />
-         </font>
-      </Text>
       <TextField fx:id="fieldIdade" layoutX="127.0" layoutY="265.0" prefHeight="25.0" prefWidth="89.0" />
       <Separator layoutX="57.0" layoutY="346.0" prefHeight="6.0" prefWidth="327.0" />
-      <Text fx:id="textSenha" layoutX="59.0" layoutY="331.0" strokeWidth="0.0" text="Senha: ">
-         <font>
-            <Font size="16.0" />
-         </font>
-      </Text>
       <Separator layoutX="57.0" layoutY="185.0" prefHeight="6.0" prefWidth="327.0" />
-      <Text fx:id="textNome" layoutX="57.0" layoutY="170.0" strokeWidth="0.0" text="Nome: ">
-         <font>
-            <Font size="16.0" />
-         </font>
-      </Text>
       <TextField fx:id="filedNome" layoutX="125.0" layoutY="151.0" prefHeight="25.0" prefWidth="254.0" />
       <Button fx:id="btVoltar" layoutX="252.0" layoutY="424.0" mnemonicParsing="false" onAction="#voltar" text="Voltar">
          <font>
@@ -59,11 +29,41 @@
       </Button>
       <PasswordField fx:id="fieldSenha" layoutX="125.0" layoutY="312.0" prefHeight="25.0" prefWidth="89.0" />
       <Separator layoutX="57.0" layoutY="397.0" prefHeight="6.0" prefWidth="327.0" />
-      <Text fx:id="textExpecialidade" layoutX="59.0" layoutY="382.0" strokeWidth="0.0" text="Expecialidade:  ">
-         <font>
-            <Font size="16.0" />
-         </font>
-      </Text>
       <PasswordField fx:id="fieldExpecialidade" layoutX="176.0" layoutY="363.0" prefHeight="25.0" prefWidth="198.0" />
+      <Label fx:id="label1" layoutX="121.0" layoutY="32.0" text="Cadastrar Usuários">
+         <font>
+            <Font size="24.0" />
+         </font>
+      </Label>
+      <Label fx:id="label2" layoutX="74.0" layoutY="86.0" text="Médico">
+         <font>
+            <Font size="18.0" />
+         </font>
+      </Label>
+      <Label fx:id="label3" layoutX="56.0" layoutY="150.0" text="Nome:">
+         <font>
+            <Font size="18.0" />
+         </font>
+      </Label>
+      <Label fx:id="label4" layoutX="63.0" layoutY="206.0" text="CPF: ">
+         <font>
+            <Font size="18.0" />
+         </font>
+      </Label>
+      <Label fx:id="label5" layoutX="57.0" layoutY="264.0" text="Idade: ">
+         <font>
+            <Font size="18.0" />
+         </font>
+      </Label>
+      <Label fx:id="label6" layoutX="54.0" layoutY="311.0" text="Senha: ">
+         <font>
+            <Font size="18.0" />
+         </font>
+      </Label>
+      <Label fx:id="label7" layoutX="51.0" layoutY="362.0" text="Expecialidade:  ">
+         <font>
+            <Font size="18.0" />
+         </font>
+      </Label>
    </children>
 </AnchorPane>
