+++ src/gui/guifx/TelaMarcarExame.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
-<AnchorPane id="AnchorPane" prefHeight="500.0" prefWidth="440.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="gui.guifx.FXMLTelaInicialController">
+<AnchorPane id="AnchorPane" prefHeight="500.0" prefWidth="440.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="gui.guifx.TelaMarcarExameController">
    <children>
       <Button fx:id="btVoltar" layoutX="240.0" layoutY="421.0" mnemonicParsing="false" onAction="#botaoVoltar" text="Voltar">
          <font>
