+++ src/gui/guifx/TelaMarcarExameController.java
@@ -25,6 +25,7 @@ public class TelaMarcarExameController implements Initializable {
      * Initializes the controller class.
      */
     
+    
     @FXML
     private Button btVoltar;
